Move MapillaryLayer to appropriate package
