+++ src/main/java/org/openstreetmap/josm/plugins/mapillary/MapillaryPlugin.java
@@ -40,6 +40,7 @@ import org.openstreetmap.josm.plugins.mapillary.gui.dialog.MapillaryFilterDialog
 import org.openstreetmap.josm.plugins.mapillary.gui.dialog.MapillaryHistoryDialog;
 import org.openstreetmap.josm.plugins.mapillary.gui.imageinfo.ImageInfoHelpPopup;
 import org.openstreetmap.josm.plugins.mapillary.gui.imageinfo.ImageInfoPanel;
+import org.openstreetmap.josm.plugins.mapillary.gui.layer.MapillaryLayer;
 import org.openstreetmap.josm.plugins.mapillary.gui.layer.PointObjectLayer;
 import org.openstreetmap.josm.plugins.mapillary.oauth.MapillaryUser;
 import org.openstreetmap.josm.plugins.mapillary.utils.MapillaryProperties;
