+++ src/main/java/org/openstreetmap/josm/plugins/mapillary/actions/package-info.java
@@ -1,7 +1,7 @@
 // License: GPL. For details, see LICENSE file.
 /**
  * Actions that are normally attached to a button or a menu item and are then executed
- * to manipulate the {@link org.openstreetmap.josm.plugins.mapillary.MapillaryLayer} when
+ * to manipulate the {@link org.openstreetmap.josm.plugins.mapillary.gui.layer.MapillaryLayer} when
  * the button is clicked.
  */
 package org.openstreetmap.josm.plugins.mapillary.actions;
