+++ src/main/java/org/openstreetmap/josm/plugins/mapillary/history/commands/CommandImport.java
@@ -6,7 +6,7 @@ import static org.openstreetmap.josm.tools.I18n.trn;
 import java.util.Set;
 
 import org.openstreetmap.josm.plugins.mapillary.MapillaryAbstractImage;
-import org.openstreetmap.josm.plugins.mapillary.MapillaryLayer;
+import org.openstreetmap.josm.plugins.mapillary.gui.layer.MapillaryLayer;
 
 /**
  * Imports a set of images stored locally.
